% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{scatterplots}
\alias{scatterplots}
\title{scatterplots}
\usage{
scatterplots(file, directory = getwd())
}
\arguments{
\item{expr.matrix}{the raw or normlaised count matrix}
}
\value{
plots
}
\description{
Generates the pairwise correlation between the samples
}
\examples{

}
